falcon fixed fmash.c to delcare final value variables as 32 signed int instead of 8 bit unsigned
git-svn-id: svn+ssh://136.177.121.23/svn/asldev/q330asp@145 7e50f963-adc0-43f8-a874-3926c70588ee

+++ trunk/falcon/fmash.c
@@ -351,9 +351,9 @@ int fmash_msh_to_csv( csv_buffer_t** csv, uint8_t* raw_msh, size_t length )
     int32_t last_high = 0;
     int32_t last_low = 0;
 
-    uint8_t final_average = 0;
-    uint8_t final_high = 0;
-    uint8_t final_low = 0;
+    int32_t final_average = 0;
+    int32_t final_high = 0;
+    int32_t final_low = 0;
 
     uint32_t temp_uint32 = 0;
     uint16_t temp_uint16 = 0;
